HBX-1999: Remove class 'org.hibernate.tool.internal.reveng.DefaultAssociationInfo' and implement factory method for AssociationInfo in RevengUtils
